netwerk/test/mochitests/test_1503201.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1503201
-->
<head>
  <title>A WWW-Authenticate response header with an invalid realm doesn't crash the browser</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1503201">Mozilla Bug 1503201</a>
<p id="display"></p>
<div id="content" style="display: none">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();
fetch("file_1503201.sjs")
  .then(() => ok(true, "no crash"))
  .then(() => SimpleTest.finish());

</script>

</div>
<pre id="test">
</pre>
</body>
</html>
